refactor: error 패키지로 이동 및 기존 import 수정
