+++ src/main/java/com/board/core/exception/AuthFailedException.java
@@ -1,5 +1,7 @@
 package com.board.core.exception;
 
+import com.board.core.exception.error.ErrorCode;
+
 public class AuthFailedException extends CustomException{
     public AuthFailedException(ErrorCode errorCode) {
         super(errorCode);
