+++ src/main/java/com/board/core/exception/CustomException.java
@@ -1,5 +1,7 @@
 package com.board.core.exception;
 
+import com.board.core.exception.error.ErrorCode;
+
 public class CustomException extends RuntimeException{
     private final ErrorCode errorCode;
