+++ src/main/java/com/board/core/exception/GlobalExceptionHandler.java
@@ -1,5 +1,6 @@
 package com.board.core.exception;
 
+import com.board.core.exception.error.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
